Conflicts:
	debian/control
	debian/control.in

+++ raster/r.terraflow/main.cc
@@ -175,7 +175,7 @@ parse_args(int argc, char *argv[]) {
   if (strcmp(d8cut->answer, "infinity") == 0) {
     opt->d8cut = MAX_ACCU;
   } else {
-    opt->d8cut = atoi(d8cut->answer);
+    opt->d8cut = atof(d8cut->answer);
   }
 
   opt->mem = atoi(mem->answer);
@@ -339,7 +339,7 @@ void record_args(int argc, char **argv) {
   sprintf(buf, "D8CUT=%f", opt->d8cut);
   stats->comment(buf);
 
-  size_t mm_size = opt->mem  << 20; /* (in bytes) */
+  size_t mm_size = (size_t) opt->mem  << 20; /* (in bytes) */
   char tmp[100];
   formatNumber(tmp, mm_size);
   sprintf(buf, "Memory size: %s bytes", tmp);
@@ -529,7 +529,7 @@ main(int argc, char *argv[]) {
   }
 
   /* set up STREAM memory manager */
-  size_t mm_size = opt->mem << 20; /* opt->mem is in MB */
+  size_t mm_size = (size_t) opt->mem << 20; /* opt->mem is in MB */
   MM_manager.set_memory_limit(mm_size);
   if (opt->verbose) {
 	MM_manager.warn_memory_limit();
